move sdk::feature utils into feature-gate-interface
